Refactor: Moved all map logic to its own class
