+++ lib/map.dart
@@ -1,0 +1,91 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_map/flutter_map.dart';
+import 'package:geolocator/geolocator.dart';
+import 'package:latlong2/latlong.dart';
+
+class OpenStreetMap extends StatefulWidget {
+  const OpenStreetMap({super.key, required this.title});
+  final String title;
+
+  @override
+  State<OpenStreetMap> createState() => _OpenStreetMapState();
+}
+
+class _OpenStreetMapState extends State<OpenStreetMap> {
+  final MapController _mapController = MapController();
+  LatLng _currentPosition = LatLng(53.2194, 6.5665); // Default to Groningen
+
+  @override
+  void initState() {
+    super.initState();
+    _determinePosition();
+  }
+
+  Future<void> _determinePosition() async {
+    bool serviceEnabled;
+    LocationPermission permission;
+
+    serviceEnabled = await Geolocator.isLocationServiceEnabled();
+    if (!serviceEnabled) {
+      print('Location services are disabled.');
+      return;
+    }
+
+    permission = await Geolocator.checkPermission();
+    if (permission == LocationPermission.denied) {
+      permission = await Geolocator.requestPermission();
+      if (permission == LocationPermission.denied) {
+        print('Location permissions are denied');
+        return;
+      }
+    }
+
+    if (permission == LocationPermission.deniedForever) {
+      print('Location permissions are permanently denied.');
+      return;
+    }
+
+    final Position position = await Geolocator.getCurrentPosition();
+    final newPosition = LatLng(position.latitude, position.longitude);
+
+    setState(() {
+      _currentPosition = newPosition;
+    });
+
+    _mapController.move(newPosition, _mapController.camera.zoom);
+
+    print('Current location: $newPosition');
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text(widget.title)),
+      body: FlutterMap(
+        mapController: _mapController,
+        options: MapOptions(initialCenter: _currentPosition, initialZoom: 13),
+        children: [
+          TileLayer(
+            urlTemplate: 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+            subdomains: const ['a', 'b', 'c'],
+            userAgentPackageName: 'com.example',
+          ),
+          MarkerLayer(
+            markers: [
+              Marker(
+                point: _currentPosition,
+                width: 40,
+                height: 40,
+                child: const Icon(
+                  Icons.location_pin,
+                  color: Colors.red,
+                  size: 40,
+                ),
+              ), // TODO: Test marker, need to remove later
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
